test(contracts): fix tx note emitter tests
This branch updates the tx types that are used inside the
`TransactionNoteEmitter` unit tests to be as same as the types that have
been defined in the `TransactionNoteEmitter` contract.

+++ contracts/contracts/protocol/v1/mocks/MockTransactionNoteEmitter.sol
@@ -9,7 +9,7 @@ contract MockTransactionNoteEmitter is TransactionNoteEmitter {
 
     function internalSanitizePrivateMessage(
         bytes memory privateMessages,
-        uint8 txType
+        uint16 txType
     ) external {
         _sanitizePrivateMessage(privateMessages, txType);
